refactor(pipeline): implement Scikit-Learn/Imblearn pipelines and optimize thresholds - feature_eng: migrate preprocessing to Scikit-Learn ColumnTransformer. - models: implement Imblearn Pipeline to conditionally apply SMOTE. - models: replace hard .predict() with .predict_proba() to enable custom thresholds. - models: tune model hyperparameters inside _get_winner_model.

+++ src/churn_analysis/etl/_schemas.py
@@ -16,7 +16,7 @@ class CustomerRecord(BaseModel):
     senior_citizen: bool = Field(alias="SeniorCitizen")
     partner: bool = Field(alias="Partner")
     dependents: bool = Field(alias="Dependents")
-    tenure: int = Field(ge=0)
+    tenure: float = Field(ge=0)
     phone_service: bool = Field(alias="PhoneService")
     multiple_lines: str = Field(alias="MultipleLines")
     internet_services: str = Field(alias="InternetService")
